requirements.txt
sentence-transformers
langchain_community
pypdf
langchain-google-genai
langchain-pinecone
python-dotenv
langchain-huggingface
torch